ballin

+++ src/main/java/e2p/icotp/service/server/dao/LoanTypeDAO.java
@@ -52,14 +52,14 @@ public class LoanTypeDAO {
     }
 
     public static void updateById(LoanType loan_type, long target_id) {
-        SQLParam idParam = new SQLParam(Types.BIGINT, "type_id", target_id);
+        SQLParam idParam = new SQLParam(Types.INTEGER, "type_id", target_id);
         ArrayList<SQLParam> params = parameters(loan_type);
         SQLCommand.updateById("loan_types", params, idParam);
     }
 
     // Remove
     public static void remove(int type_id) {
-        SQLParam idParam = new SQLParam(Types.BIGINT, "type_id", type_id);
+        SQLParam idParam = new SQLParam(Types.INTEGER, "type_id", type_id);
         SQLCommand.deleteById("loan_types", idParam);
     }
 
@@ -68,7 +68,7 @@ public class LoanTypeDAO {
         ArrayList<SQLParam> params = new ArrayList<>();
 
         // int type_id
-        params.add(new SQLParam(Types.BIGINT, "type_id", loan_type.getId().get()));
+        params.add(new SQLParam(Types.INTEGER, "type_id", loan_type.getId().get()));
         // varchar type_name
         params.add(new SQLParam(Types.VARCHAR, "type_name", loan_type.getName().get()));
         // varchar type_desc
